Added a player asset and animations
Right now the character have 2 animations, idle and walk. Will add more in the future.
