+++ Assets/AnimationController.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7f0be64f6d7314944b960fa6ba98c125
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
