+++ Assets/Player1.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7b53b06ac6853da459b6a68f69a6a3da
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
