+++ Assets/Player1/Eyebrow.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 95bc9feddd5d2aa409016a1640e1cc0e
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
