+++ Assets/Player1/Hair.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fc6aa867acfcd0d4ab787bbda8063dbf
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
