+++ Assets/Player1/Idle.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a1accb3b79cd35943ae82bdd93bfd2d7
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
