+++ Assets/Player1/Shirt.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e5842324ff023294c9e1950237cc393c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
