+++ Assets/Player1/Shoes.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 172888a762d52e047b7834438d493dbf
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
