+++ Assets/Player1/Shorts.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2b02085ba75764b4892eb0e553b71ef0
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
