+++ Assets/Player1/Skin.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f0ccc433821f427428689cb5f20c2220
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
