+++ Assets/Player1/Walk.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5e6df00f91535a349a8a06af593a6cf0
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
