+++ Assets/Scripts/AnimationController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 96b61178310c2b54f84ef16798ec3019
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
